chore: rename binary_gender to sex
